intelligent_observatory/requirements.txt
flask
flask-login
flask-sqlalchemy
mysqlclient
pandas
scikit-learn
requests
beautifulsoup4
python-dotenv
matplotlib
seaborn